+++ lib/hux.dart
@@ -56,6 +56,7 @@ export 'src/components/feedback/hux_alert.dart';
 export 'src/components/feedback/hux_snackbar.dart';
 export 'src/components/avatar/hux_avatar.dart';
 export 'src/components/avatar/hux_avatar_group.dart';
+export 'src/components/tabs/hux_tabs.dart';
 export 'src/components/tooltip/hux_tooltip.dart';
 export 'src/theme/hux_theme.dart';
 export 'src/theme/hux_colors.dart';